Reorganize renet into client-plugin and server
